split into separate impl and file structure.
