+++ src/main.rs
@@ -1,34 +1,11 @@
-use std::io::{stdin, stdout, Read};
-use termion::raw::IntoRawMode;
+mod editor;
 
-fn to_cntrl_byte(c: char) -> u8 {
-    let byte = c as u8;
-    byte & 0b0001_1111
-}
-
-fn die(e: std::io::Error) {
-    panic!("{}", e);
-}
+use editor::Editor;
 
 fn main() {
-    println!("Starting Rustypo, Enter Text...");
+    println!("Starting Rustypo, Enter Text...");  
 
-    let _stdout = stdout().into_raw_mode().unwrap();
-
-    for b in stdin().bytes() {
-        match b {
-            Ok(b) => {
-                let c = b as char;
-                if c.is_control() {
-                    println!("{:?} \r", b); 
-                } else {
-                    println!("{:?} ({})\r", b, c); 
-                }
-                if b == to_cntrl_byte('c'){ //cant use ctrl q - is a hotkey on my ide!
-                    break;
-                }
-            }
-            Err(err) => die(err),
-        } 
-    }
+    let editor = Editor {};
+    editor.run();
 }
+
